# dist/api.yml
openapi: 3.0.0
info:
  title: api
  description: API for  project
  version: 1.0.0
...
paths:
  /Register:
    post:
      tags:
        - Auth
      summary: Add new user
      description: A new user from the client is added to the server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
  /Login:
    post:
      tags:
        - Auth
      summary: Login an exist user
      description: An exist user login to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginScheme"
      responses:
        "200":
          description: >
            User was found in the DB and authenitcated successfully.
            The session ID is returned in a cookie named `session`. 
            You should include this cookie in subsequent requests
          # headers:
          #   Set-Cookie:
          #     schema:
          #       type: string
          #       example: session=87s97sdhgf279293f8w0e9
            # Note that the Set-Cookie header and securitySchemes are not connected in any way, and the Set-Header definition is for documentation purposes only.
        "401":
          description: Authentication failed
...
  /manager/addGame:
    post:
      tags:
        - manager
      summary: A union manager from client is add game to league
      description: a JSON object contains gameId to add to league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addgame'
        description: add new game
      responses:
        '201':
          description: game has been added
        '400':
          description: incorrect inputs
        '403':
          description: The user doesnt have access to add game
  /manager/addEvent:
    post:
      tags:
        - manager
      summary: A union manager from client is add game to league
      description: a JSON object contains gameId to add to league
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/eventslog'
        description: add new event
      responses:
        '201':
          description: event has been added
        '400':
          description: incorrect inputs
        '403':
          description: The user doesnt have access to add game
  /manager/addScore:
    put:
      tags:
        - manager
      summary: A union manager from client is add score to game
      description: a JSON object contains score to add to game
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                gameId:
                  type: int
                  description: game id which the user saved
                  example: 13
                homegoal:
                  type: int
                  description: home goal
                  example: 2
                awaygoal:
                  type: int
                  description: away goal
                  example: 1
      responses:
        '200':
          description: score has been added to game
        '400':
          description: incorrect inputs
        '403':
          description: The user doesnt have access to add game
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins between 3 to 8 letters only
          example: ladygaga
        firstname:
          type: string
          description: user's first name
          example: Stefani
        lastname:
          type: string
          description: user's last name
          example: Germanotta
        country:
          type: string
          description: user's country
          example: USA
        password:
          maxLength: 10
          minLength: 5
          type: string
          description:
            password conatins between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: lady@56
        email:
          pattern: email
          type: string
          description: user's email
          example: ladygaga@gmail.com
        imageUrl:
          pattern: URL
          type: string
          description: URL for the user's image
          example: https://cloudinary.com/
      description: Model containing user info
    LoginScheme:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          description: username conatins letters only
          example: ladygaga
        password:
          type: string
          minLength: 5
          maxLength: 10
          format: password
          description: password conatins at least one number and one special char
          example: lady@56
...
    addgame:
      type: object
      description: Contains data on new game
      properties:
        gamedate:
          type: string
          example: '2022-12-10'
          description: the date of the game
        gametime:
          type: string
          example: '20:45:00'
          description: the time of the game
        hometeamID:
          type: integer
          example: 86
          description: id of home team
        awayteamID:
          type: integer
          example: 85
          description: id of away team
        field:
          type: string
          example: Parken
          description: field
        referee:
          type: string
          example: Andrew Dallas
          description: referee
    eventslog:
      type: object
      description: Contains game events
      properties:
        gameID:
          type: integer
          example: 10
          description: the id of the game
        eventminute:
          type: integer
          example: 10
          description: the minute of the game
        dataevent:
          type: string
          example: goal
          description: event
        playerID:
          type: integer
          example: 82808
          description: the id of the player who did the event